migrate long job notification to snack-bar
